app/requirements.txt
Click==7.0
Flask==1.1.1
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.1
joblib==0.14.1
MarkupSafe==1.1.1
numpy==1.18.1
pandas==1.0.1
python-dateutil==2.8.1
pytz==2019.3
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
Werkzeug==1.0.0
WTForms==2.2.1
